Create new explicit Install command and remove check for registered alias on every run
